% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base_plot.R
\docType{data}
\name{base_pressure_plot}
\alias{base_pressure_plot}
\title{Base pressure chart}
\format{An object of class \code{gg} (inherits from \code{ggplot}) of length 9.}
\usage{
base_pressure_plot
}
\description{
This chart corresponds to the printed chart, updated for
current tire sizes. The \code{\link{plot_bike_inflation}} functions superimpose
the front and rear calculated pressures for a bike setup.
}
\keyword{datasets}
