% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-tire-pressures.R
\name{bike_tire_pressures}
\alias{bike_tire_pressures}
\title{Compute front and rear tire pressures based on data for bike and rider.}
\usage{
bike_tire_pressures(rider_weight_lbs = 100, bike_weight_lbs = 15,
  load_lbs = 2, front_tire_casing_compensation = 1,
  front_tire_size_mm = 28, rear_tire_casing_compensation = 1,
  rear_tire_size_mm = 28, front_distribution = 0.5)
}
\arguments{
\item{rider_weight_lbs}{Weight of rider dressed with biking shoes and helmet.}

\item{bike_weight_lbs}{Weight of bike}

\item{load_lbs}{Weight of accessories carried on the bike such as water bottles, repair kit, pump, etc.}

\item{front_tire_casing_compensation}{Supple casings require about 15\% more pressure to maintain shape.}

\item{front_tire_size_mm}{Measured or side wall marking.}

\item{rear_tire_casing_compensation}{Supple casings require about 15\% more pressure to maintain shape.}

\item{rear_tire_size_mm}{Measured or side wall marking.}

\item{front_distribution}{Decimal representation of percentage of bike plus rider weight on the front wheel.}
}
\value{
data.frame that used by ggplot to display tire pressures
}
\description{
Compute front and rear tire pressures based on data for bike and rider.
}
