% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btpress-package.R
\docType{package}
\name{btpress}
\alias{btpress}
\alias{btpress-package}
\title{btpress}
\description{
If you were told that you could bicycle faster and more comfortablely simply
by changing your bike tire pressure, would you try it? There are now 2
independent studies that support lower pressures are significantly better
than pumping up your tires until they are rock hard.  See the \link{#References} for links.
}
\section{Pump Inccuracies}{


Pump gauges considerably inaccurate. Parallex viewing makes precise reading
of needle impossible.
}

\section{Weight vs. Load}{


At the end of the day, these two are effectively the same.  That said,
\emph{Weight} is used for the individual components that go into the calculations.
\emph{Load} is used
to describe the forces the tires are subject to.
}

\section{Directions for Use}{


Extra light casings need an extra 10-15\% pressure to prevent the casing threads from breaking.
}

\section{References}{

Links to articles:

\itemize{
  \item fill in
  \item fill in
  \item \href{https://www.compasscycle.com/wp-content/uploads/2015/03/BQTireDrop.pdf}{Bicycle Quarterly article}
  \item \href{https://www.hedcycling.com/blog/road-tire-pressure-for-plus-and-black-wheels/}{Road Tire Pressure For Plus And BLACK Wheels}
}
Bicycle Quarterly blogs
Silca.cc - Josh's articles
}

\section{Acknowledgements}{


Source for data and pratical application to Jan Heine and Frank Berto
published in \href{https://www.compasscycle.com/wp-content/uploads/2015/03/BQTireDrop.pdf}{Bicycle Quarterly}.

The emperical curve fitting formula from the
\href{http://www.biketinker.com/2010/bike-resources/optimal-tire-pressure-for-bicycles/}{BikeTinker}.
}

\section{Full Disclosure}{


I ride and recommend Compass Bicycle
\href{https://www.compasscycle.com/product-category/components/tires/}{tires}
and subscribe to \href{https://www.bikequarterly.com/}{\emph{Bicycle
Quarterly}}. I received no compensation for this program nor was I asked to
write it by anyone.
}

\section{Disclaimer}{


Biking riding is an inheritly risky activity. Remember the bike racer's motto: \emph{Falling is a matter of
'when', not 'if'.} You are solely responsible for not extending any load limits
including, but not
exclusive of maximum tire inflation pressure, rim inflation pressure,
mininum tire clearance, and
compatibility of tubeless tire and rim combinations.
Use of this app is at your risk and you assume all liabilities.
}

