% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike-tire-pressures.R
\name{check_pressure}
\alias{check_pressure}
\title{Check if pressure exceeds certain levels}
\usage{
check_pressure(p, warn_psi = 105, max_psi = 120)
}
\arguments{
\item{p}{Desired tire inflation pressure}

\item{warn_psi}{Pressure exceeds to suggest increasing tire size}

\item{max_psi}{Pressure exceeds warn to check max pressure ratings of tire & rim}
}
\value{
a list or NA
}
\description{
More details here...
}
\examples{
Fill in later....
}
