% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-bikes.R
\name{create_example_bikes}
\alias{create_example_bikes}
\title{Create a few example bikes and inflation plots}
\usage{
create_example_bikes()
}
\value{
creates bike and plot objects in the Global environment
}
\description{
Create a few example bikes and inflation plots
}
\examples{
\donttest{
\dontrun{
create_example_bikes()
steel_sport
steel_sport_plot
}
}
}
