% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{display_bike_inflation}
\alias{display_bike_inflation}
\title{Plot and label front and rear wheel inflation data for a bike.}
\usage{
display_bike_inflation(base_plot = base_pressure_plot, bike)
}
\arguments{
\item{base_plot}{ggplot object of base tire pressure curves.}

\item{bike}{data.frame to display for specific bike and rider.}
}
\value{
A complete plot for display
}
\description{
Plot and label front and rear wheel inflation data for a bike.
}
