% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droop_pressure.R
\name{droop_pressure_psi}
\alias{droop_pressure_psi}
\title{Compute the 15\% droop tire pressure based on wheel load and tire size}
\usage{
droop_pressure_psi(weight_lbs, tire_size_mm)
}
\arguments{
\item{weight_lbs}{is the load on the bicycle wheel and typically 40 - 60
percent of the total weight of the rider, bike and carried items.}

\item{tire_size_mm}{is the nominal size of the tire from the casing label or
directly measured diameter.}
}
\value{
pressure, psi
}
\description{
Original research used rims with internal widths of ~19mm. Today's wider
rims can add increase the volume of a tire by 33\%.
}
\details{
The emperical, imperial centric formula is from the
\href{http://www.biketinker.com/2010/bike-resources/optimal-tire-pressure-for-bicycles/}{BikeTinker}
}
