% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{dual_pressure}
\alias{dual_pressure}
\title{Y axis label formatting function.}
\usage{
dual_pressure(psi)
}
\arguments{
\item{psi}{}
}
\value{
Formatted string with psi and bars.
}
\description{
Used for labeling plot axis
}
