% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{dual_pressure_point}
\alias{dual_pressure_point}
\title{Data point formatting function.}
\usage{
dual_pressure_point(tire_size, position = c("Front", "Rear"), psi)
}
\arguments{
\item{tire_size}{String in the form "<integer>mm".}

\item{position}{Indicates which wheel. Value should be "Front" or "Rear".}

\item{psi}{Inflation pressure for the wheel.}
}
\value{
string similar to
\preformatted{
    28mm Front
 80 psi | 5.5 bar
}
}
\description{
Data point formatting function.
}
