% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{dual_weight}
\alias{dual_weight}
\title{X axis label formatting function}
\usage{
dual_weight(lbs)
}
\arguments{
\item{lbs}{weight in pounds.}
}
\value{
Formatted string with pounds and kilograms.
}
\description{
X axis label formatting function
}
