% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base_plot.R
\name{generate_base_pressure_plot}
\alias{generate_base_pressure_plot}
\title{Generate the base plot of inflation curves for all tire sizes}
\usage{
generate_base_pressure_plot(base_inflation = generate_inflation_data(),
  plot_theme = theme_dg_pale())
}
\arguments{
\item{base_inflation}{the inflation data to use.}

\item{plot_theme}{the ggplot theme to use.}
}
\value{
ggplot object
}
\description{
Generate the base plot of inflation curves for all tire sizes
}
