% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-inflation-curve-data.R
\name{generate_inflation_data}
\alias{generate_inflation_data}
\title{Generate the dataset for drawing the inflation curves}
\usage{
generate_inflation_data(wheel_loads = wheel_loads_lbs(),
  tire_sizes = tire_sizes_mm(), min_psi = 15, max_psi = 120)
}
\arguments{
\item{wheel_loads}{Vector containing the major X axis loads.}

\item{tire_sizes}{Vector containing the different tire widths.}

\item{min_psi}{Lowest pressure to plot.}

\item{max_psi}{Highest pressure to plot.}
}
\value{
tibble
}
\description{
Assign it to a variable if you want to explore it.
}
