% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{heavy_rider_expand_limits}
\alias{heavy_rider_expand_limits}
\title{heavy_rider_expand_limits}
\usage{
heavy_rider_expand_limits(bike, plot)
}
\arguments{
\item{bike}{setup to be plotted}

\item{plot}{base_plot}
}
\value{
ggplot object
}
\description{
heavy_rider_expand_limits
}
