% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-inflation-curve-data.R
\name{inflation_datum}
\alias{inflation_datum}
\title{Generate a data set point.  Used to build data to display inflation curves on
base inflation plot.}
\usage{
inflation_datum(wheel_load_lbs, tire_size_mm)
}
\arguments{
\item{wheel_load_lbs}{Self-explanatory}

\item{tire_size_mm}{Self-explanatory}
}
\value{
Vector of wheel load, tire size, and droop pressure
}
\description{
Generate a data set point.  Used to build data to display inflation curves on
base inflation plot.
}
