% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-bike-inflation.R
\name{plot_bike_inflation}
\alias{plot_bike_inflation}
\title{Plot and label front and rear wheel inflation data for a bike setup}
\usage{
plot_bike_inflation(base_plot = base_pressure_plot, bike, show_note = FALSE,
  note = NULL)
}
\arguments{
\item{base_plot}{ggplot object of base tire pressure curves.}

\item{bike}{tibble to display for specific bike and rider.}

\item{show_note}{boolean to show the note under the title.}

\item{note}{string to print under title}
}
\value{
ggplot object with the bike superimposed over the base inflation plot.
}
\description{
Plot and label front and rear wheel inflation data for a bike setup
}
