% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{plot_title}
\alias{plot_title}
\title{Display a generic title and optionally add bike summary info}
\usage{

  plot_title(title = "Optimized Bicycle Tire Pressure for Road & Gravel Riding",
  subtitle = NULL)
}
\arguments{
\item{title}{for the plot.}

\item{subtitle}{If provided, is used for the subtitle.}
}
\value{
labels
}
\description{
Wrapper for \code{\link[ggplot2]{ggtitle}} so you can change the
the title and subtitle for your application.
}
