% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{psi_to_bar}
\alias{psi_to_bar}
\title{Convert PSI to Bar}
\usage{
psi_to_bar(psi)
}
\arguments{
\item{psi}{Pressure in pounds per square inch.}
}
\value{
bars
}
\description{
1 Psi = 0.0689475729 Bar
}
