% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\docType{data}
\name{recommended_pressure_color}
\alias{recommended_pressure_color}
\title{Better safety and comfort color}
\format{An object of class \code{character} of length 1.}
\usage{
recommended_pressure_color
}
\description{
Better safety and comfort color
}
\keyword{internal}
