% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{rider_expand_limits}
\alias{rider_expand_limits}
\title{rider_expand_limits}
\usage{
rider_expand_limits(bike)
}
\arguments{
\item{bike}{setup to be plotted}
}
\value{
ggplot object
}
\description{
Wheel loads at extremes need to expand the plot area to prevent
labels from being clipped.
}
