% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dg_pale.R
\name{theme_dg_pale}
\alias{theme_dg_pale}
\title{Modifications to \code{theme_bw} to add a bit of color}
\usage{
theme_dg_pale(...)
}
\arguments{
\item{...}{passed to theme_bw}
}
\value{
ggplot theme
}
\description{
Modifications to \code{theme_bw} to add a bit of color
}
\seealso{
\code{\link[ggplot2]{ggtheme}}
}
