% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-inflation-curve-data.R
\docType{data}
\name{tire_sizes_mm}
\alias{tire_sizes_mm}
\title{Nominal tire sizes.}
\format{An object of class \code{integer} of length 10.}
\usage{
tire_sizes_mm
}
\description{
Also happen to be the ones Compass Bicycle sells.
}
\details{
14-Jun-2017: Updated to include all sizes.
}
\keyword{datasets}
