% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-inflation-curve-data.R
\docType{data}
\name{wheel_loads_lbs}
\alias{wheel_loads_lbs}
\title{X axis values for wheel loads in lbs}
\format{An object of class \code{numeric} of length 9.}
\usage{
wheel_loads_lbs
}
\description{
Based on the Bicycle Quarterly chart which is in kilograms
}
\keyword{datasets}
