% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike-tire-pressures.R
\name{check_pressure}
\alias{check_pressure}
\title{Check if pressure exceeds certain levels}
\usage{
check_pressure(p, tire_size)
}
\arguments{
\item{p}{Desired tire inflation pressure}

\item{tire_size}{tire size}
}
\value{
a list or NA
}
\description{
Display a message if the tire pressure might be getting into the range where
a tire or rim could be overinflated. Data published by DT Swiss listing the
maxiumum rim pressure by tire size is used.
}
