% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-bikes.R
\name{create_example_bikes}
\alias{create_example_bikes}
\title{Create a few example bikes and inflation plots}
\usage{
create_example_bikes()
}
\value{
list of bikes
}
\description{
Create a few example bikes and inflation plots
}
