% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{generate_base_pressure_plot}
\alias{generate_base_pressure_plot}
\title{Generate the base plot of inflation curves for all tire sizes.}
\usage{
generate_base_pressure_plot(data = inflation_data, plot_theme = NA)
}
\arguments{
\item{data}{the inflation data to use}

\item{plot_theme}{the ggplot theme to use for the plot.  Currently
not implemented.}
}
\value{
ggplot object
}
\description{
Generate the base plot of inflation curves for all tire sizes.
}
