% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{generate_base_pressure_plot}
\alias{generate_base_pressure_plot}
\title{#' ggplot theme tweaks to improve aesthics in a Shiny app.
#'
#' Starts with dgutils::theme_dg if found; othewise theme_bw.
#'
#' @param base_theme Theme to use instead of theme_dg or theme_bw.
#'
#' @return Theme object.
#' @export
theme_dg_shiny <- function(base_theme) {
  if (is.null(base_theme) | is.na(base_theme)) {
    if (exists("theme_dg",  where = "package:dgutils")) { theme_dg } else { theme_bw() }
  } else { base_theme } +
    theme_update(
      plot.subtitle = element_text(size = rel(1.1)),
      plot.caption = element_text(color = "#cccccc"),
      axis.title = element_text(face = "bold"),
      axis.text = element_text(size = rel(1.05)),
      aspect.ratio = 0.66,
      legend.position = "none" # Avoid show.legend = "FALSE" args
    )
}}
\usage{
generate_base_pressure_plot(data = inflation_data, plot_theme = NA)
}
\description{
#' ggplot theme tweaks to improve aesthics in a Shiny app.
#'
#' Starts with dgutils::theme_dg if found; othewise theme_bw.
#'
#' @param base_theme Theme to use instead of theme_dg or theme_bw.
#'
#' @return Theme object.
#' @export
theme_dg_shiny <- function(base_theme) {
  if (is.null(base_theme) | is.na(base_theme)) {
    if (exists("theme_dg",  where = "package:dgutils")) { theme_dg } else { theme_bw() }
  } else { base_theme } +
    theme_update(
      plot.subtitle = element_text(size = rel(1.1)),
      plot.caption = element_text(color = "#cccccc"),
      axis.title = element_text(face = "bold"),
      axis.text = element_text(size = rel(1.05)),
      aspect.ratio = 0.66,
      legend.position = "none" # Avoid show.legend = "FALSE" args
    )
}
}
