% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pressure.R
\name{plot_title}
\alias{plot_title}
\title{Display a generic title or bike summary info.}
\usage{

  plot_title(title = "Optimized Bicycle Tire Pressure for 26, 650B, and 700C Sizes",
  subtitle = "For road & gravel riding", summary = NULL)
}
\arguments{
\item{title}{Fill in...}

\item{subtitle}{Fill in...}

\item{summary}{Fill in...}
}
\value{
figure out
}
\description{
Display a generic title or bike summary info.
}
\examples{
think about this one
}
