% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-inflation-curve-data.R
\docType{data}
\name{tire_sizes_mm}
\alias{tire_sizes_mm}
\title{Nominal tire sizes.}
\format{An object of class \code{integer} of length 10.}
\usage{
tire_sizes_mm
}
\description{
Widths from Compass Bicycle catalog.
}
\keyword{datasets}
