% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_sizes_loads.R
\docType{data}
\name{tire_sizes_mm}
\alias{tire_sizes_mm}
\title{Nominal tire sizes. Also happen to be the ones Compass Bicycle sells.
14-Jun-2017: Updated to include all sizes.}
\format{An object of class \code{numeric} of length 10.}
\usage{
tire_sizes_mm
}
\description{
Nominal tire sizes. Also happen to be the ones Compass Bicycle sells.
14-Jun-2017: Updated to include all sizes.
}
\keyword{datasets}
